Extract ProgressBar to wrapper that will contain all progress components
