+++ app/ui/practice/exam/timer.tsx
@@ -6,11 +6,8 @@ interface TimerProps {
 
 export default function Timer({ timeInSeconds }: TimerProps) {
     const hours = Math.floor(timeInSeconds / 3600)
-    console.log('hours: ', hours)
     const minutes = Math.floor((timeInSeconds % 3600) / 60)
-    console.log('minutes: ', minutes)
     const seconds = timeInSeconds % 60
-    console.log('seconds: ', seconds)
 
     function pad(num: number) {
         if (num < 10) return `0${num}`
